Streamline shared components, move files to clearer folders (#179)
* Move Documentation MDX components to mapping

* Move API Ref MDX components to mapping

* Use recursive string extraction on all linked headers

* Peel out NavItem to separate file

* Consolidate shared API Ref styles

* Only use documentation MDX components

* Clearer homes for Footer components

* Rehome ScrollRouter

* Remove not-actually-shared style

* Move Content to Article basic

* Move Documentation styles to clearer homes

None of these were really most directly associated with the /docs pages, which is how we use the name "Documentation"

* Remove NavTable

It't just a thin wrapper for MethodTable, and the name isn't substantially more meaningful

* Ignore package-lock.json

* Use partition in API ref wrapper
