+++ .gitignore
@@ -35,6 +35,8 @@ build/Release
 # Dependency directories
 node_modules/
 jspm_packages/
+# We use yarn, not npm
+package-lock.json
 
 # Typescript v1 declaration files
 typings/
