+++ content/api/aggregations/index.mdx
@@ -3,11 +3,11 @@ title: Aggregations
 order: 0
 ---
 
-import { NavTable } from "components/NavTable";
+import { MethodTable } from "components/MethodTable";
 
 Endpoints that aggregate data about the ledger.
 
-<NavTable title="Resource Types">
+<MethodTable title="Resource Types">
 
 |                                                      |     |
 | ---------------------------------------------------- | --- |
@@ -16,4 +16,4 @@ Endpoints that aggregate data about the ledger.
 | [Trade Aggregations](./trade-aggregations/index.mdx) |     |
 | [Fee Stats](./fee-stats/index.mdx)                   |     |
 
-</NavTable>
+</MethodTable>
