+++ content/api/errors/http-status-codes/horizon-specific/index.mdx
@@ -3,18 +3,18 @@ title: Horizon-Specific Status Codes
 order: 0
 ---
 
-import { NavTable } from "components/NavTable";
+import { MethodTable } from "components/MethodTable";
 
 These responses are specific to how a Horizon server is set up and how transactions should be formed.
 
-<NavTable title="Horizon-Specific Status Codes">
+<MethodTable title="Horizon-Specific Status Codes">
 
-|                                                              |     |
-| ------------------------------------------------------------ | --- |
-| [Transaction Failed](./transaction-failed.mdx)               |     |
-| [Transaction Malformed](./transaction-malformed.mdx)         |     |
-| [Before History](./before-history.mdx)                       |     |
-| [Stale History](./stale-history.mdx)                         |     |
-| [Timeout](./timeout.mdx)                                     |     |
+|                                                      |     |
+| ---------------------------------------------------- | --- |
+| [Transaction Failed](./transaction-failed.mdx)       |     |
+| [Transaction Malformed](./transaction-malformed.mdx) |     |
+| [Before History](./before-history.mdx)               |     |
+| [Stale History](./stale-history.mdx)                 |     |
+| [Timeout](./timeout.mdx)                             |     |
 
-</NavTable>
+</MethodTable>
