+++ content/api/errors/index.mdx
@@ -3,17 +3,17 @@ title: Errors
 order: 0
 ---
 
-import { NavTable } from "components/NavTable";
+import { MethodTable } from "components/MethodTable";
 
 After processing a request, Horizon returns a success or error response to the client. A success response will return a Status Code of 200, and an error response will return a Status Code in the range of 4XX - 5XX along with additional information about why the request could not complete successfully.
 
 There are two categories of errors: [HTTP Status Codes](./http-status-codes/index.mdx) and [Result Codes](./result-codes/index.mdx). Result Codes only follow a Transaction Failed (400) HTTP Status Code.
 
-<NavTable title="Error Categories">
+<MethodTable title="Error Categories">
 
 |  |  |
 | --- | --- |
 | [HTTP Status Codes](./http-status-codes/index.mdx) | Errors that occur at the Horizon Server level. |
 | [Result Codes](./result-codes/index.mdx) | Errors that occur at the Stellar Core level. |
 
-</NavTable>
+</MethodTable>
