+++ content/api/errors/result-codes/index.mdx
@@ -3,7 +3,7 @@ title: Result Codes
 order: 0
 ---
 
-import { NavTable } from "components/NavTable";
+import { MethodTable } from "components/MethodTable";
 
 Result Codes describe why a transaction or operation failed in Stellar Core and are communicated in the “extras” field of a Horizon response when the “Transaction Failed” Status Code is returned.
 
@@ -11,7 +11,7 @@ In the “extras” field, the errors returned are referred to as “Result Code
 
 There are three types of Result Codes: [Transaction Result Codes](./transactions.mdx), [Operation Result Codes](./operations.mdx), and [Operation-Specific Result Codes](/operation-specific/index.mdx).
 
-<NavTable title="Result Code Types">
+<MethodTable title="Result Code Types">
 
 |  |  |
 | --- | --- |
@@ -19,4 +19,4 @@ There are three types of Result Codes: [Transaction Result Codes](./transactions
 | [Operation Result Codes](./operations.mdx) | Generic errors about operation failures. |
 | [Operation-Specific Result Codes](/operation-specific/index.mdx) | Errors specific to each operation type. |
 
-</NavTable>
+</MethodTable>
