+++ content/api/errors/result-codes/operation-specific/index.mdx
@@ -3,28 +3,28 @@ title: Operation-Specific Result Codes
 order: 0
 ---
 
-import { NavTable } from "components/NavTable";
+import { MethodTable } from "components/MethodTable";
 
 These are Result Codes that communicate success (200) or failure (400) responses that are specific to each operation type.
 
 Each of Stellar's operations have unique causes of failure, and these result codes will help you diagnose and address the origins of issues.
 
-<NavTable title="Operation-Specific Result Code Types">
+<MethodTable title="Operation-Specific Result Code Types">
 
-|                                                              |     |
-| ------------------------------------------------------------ | --- |
-| [Create Account](./create-account.mdx)                       |     |
-| [Payment](./payment.mdx)                                     |     |
+|                                                                  |     |
+| ---------------------------------------------------------------- | --- |
+| [Create Account](./create-account.mdx)                           |     |
+| [Payment](./payment.mdx)                                         |     |
 | [Path Payment Strict Receive](./path-payment-strict-receive.mdx) |     |
-| [Path Payment Strict Send](./path-payment-strict-send.mdx)   |     |
-| [Manage Sell Offer](./manage-sell-offer.mdx)                 |     |
-| [Manage Buy Offer](./manage-buy-offer.mdx)                   |     |
-| [Create Passive Sell Offer](./create-passive-sell-offer.mdx) |     |
-| [Set Options](./set-options.mdx)                             |     |
-| [Change Trust](./change-trust.mdx)                           |     |
-| [Allow Trust](./allow-trust.mdx)                             |     |
-| [Account Merge](./account-merge.mdx)                         |     |
-| [Manage Data](./manage-data.mdx)                             |     |
-| [Bump Sequence](./bump-sequence.mdx)                         |     |
+| [Path Payment Strict Send](./path-payment-strict-send.mdx)       |     |
+| [Manage Sell Offer](./manage-sell-offer.mdx)                     |     |
+| [Manage Buy Offer](./manage-buy-offer.mdx)                       |     |
+| [Create Passive Sell Offer](./create-passive-sell-offer.mdx)     |     |
+| [Set Options](./set-options.mdx)                                 |     |
+| [Change Trust](./change-trust.mdx)                               |     |
+| [Allow Trust](./allow-trust.mdx)                                 |     |
+| [Account Merge](./account-merge.mdx)                             |     |
+| [Manage Data](./manage-data.mdx)                                 |     |
+| [Bump Sequence](./bump-sequence.mdx)                             |     |
 
-</NavTable>
+</MethodTable>
