+++ content/api/introduction/index.mdx
@@ -3,7 +3,7 @@ title: Horizon API Reference
 order: 0
 ---
 
-import { NavTable } from "components/NavTable";
+import { MethodTable } from "components/MethodTable";
 
 Horizon is an API for interacting with the Stellar network.
 
@@ -16,7 +16,7 @@ The Stellar Development Foundation (SDF) runs two instances of Horizon:
 - [https://horizon.stellar.org/](https://horizon.stellar.org/) for interacting with the public network
 - [https://horizon-testnet.stellar.org/](https://horizon-testnet.stellar.org/) for interacting with the [testnet](../../docs/glossary/testnet.mdx)
 
-<NavTable title="API Reference Sections">
+<MethodTable title="API Reference Sections">
 
 |  |  |
 | --- | --- |
@@ -25,4 +25,4 @@ The Stellar Development Foundation (SDF) runs two instances of Horizon:
 | [Aggregations](../aggregations/index.mdx) | Descriptions of specialized endpoints. |
 | [Errors](../errors/index.mdx) | Potential errors and what they mean. |
 
-</NavTable>
+</MethodTable>
