+++ content/api/resources/index.mdx
@@ -3,11 +3,11 @@ title: Resources
 order: 0
 ---
 
-import { NavTable } from "components/NavTable";
+import { MethodTable } from "components/MethodTable";
 
 Data on the Stellar ledger is organized according to resources. Each resource has several different endpoints.
 
-<NavTable title="Resource Types">
+<MethodTable title="Resource Types">
 
 |                                          |     |
 | ---------------------------------------- | --- |
@@ -20,4 +20,4 @@ Data on the Stellar ledger is organized according to resources. Each resource ha
 | [Trades](./trades/index.mdx)             |     |
 | [Assets](./assets/index.mdx)             |     |
 
-</NavTable>
+</MethodTable>
