+++ src/basics/Text.js
@@ -16,6 +16,10 @@ const textStyles = css`
   color: ${({ theme }) => theme.text};
 `;
 
+export const Article = styled.article`
+  position: relative;
+  margin: 0 auto;
+`;
 export const Text = styled.p`
   ${textStyles}
 `;
@@ -167,6 +171,7 @@ export const ListItem = styled.li`
 export const Small = styled.small`
   font-size: 75%;
 `;
+export const Section = styled.section``;
 export const Table = styled.table`
   display: block;
   word-break: normal;
